cart works!, separately (#106)

+++ client/store/signedInCart.js
@@ -1,12 +1,18 @@
 import axios from 'axios';
 
 const ADD_ITEM_TO_CART = 'ADD_ITEM_TO_CART';
+const ADD_ITEM_TO_GUEST_CART = 'ADD_ITEM_TO_GUEST_CART';
 const USER_ORDER = 'USER_ORDER';
 const DELETE_ITEM = 'DELETE_ITEM';
 const CHANGE_ORDER = 'CHANGE_ORDER';
 
 const addItemToCart = item => ({
   type: ADD_ITEM_TO_CART,
+  item
+});
+
+const addItemToGuestCart = item => ({
+  type: ADD_ITEM_TO_GUEST_CART,
   item
 });
 
@@ -25,11 +31,39 @@ const changeOrder = item => ({
   item
 });
 
-export const addItem = (userId, item) => dispatch =>
-  axios
-    .post(`/api/users/${userId}/order`, item)
-    .then(res => dispatch(addItemToCart(res.data)))
-    .catch(err => console.log(err));
+export const addItem = (userId, item) => dispatch => {
+  console.log(
+    'addItem: userId, item, called in signedInCart store',
+    userId,
+    item
+  );
+  //add item to localStorage
+  let prods = [];
+  prods = JSON.parse(localStorage.getItem('products'));
+  prods !== null
+    ? localStorage.setItem(
+        'products',
+        JSON.stringify(new Set(prods.concat(item.productId)))
+      )
+    : localStorage.setItem('products', JSON.stringify([item.productId]));
+  console.log(JSON.parse(localStorage.getItem('products')), 'prods after add');
+  // localStorage.setItem(
+  //   'quantity',
+  //   JSON.stringify(JSON.parse(localStorage.getItem('products')).map(i => 1)) ||
+  //     []
+  // );
+  if (!userId)
+    addItemToGuestCart(JSON.parse(localStorage.getItems('products')));
+  if (typeof userId === 'number')
+    axios
+      .post(`/api/users/${userId}/order`, item)
+      .then(res => {
+        //add item to user's order
+        dispatch(addItemToCart(res.data));
+        // console.log(userId, item, 'user adding to cart: user, prodId');
+      })
+      .catch(err => console.log(err));
+};
 
 export const getItems = userId => dispatch =>
   axios
@@ -37,31 +71,55 @@ export const getItems = userId => dispatch =>
     .then(res => dispatch(userOrder(res.data)))
     .catch(err => console.log(err));
 
-//delete user
 export const deleteTheItem = (userId, itemId) => dispatch => {
-  axios
-    .post(`/api/users/${userId}/item/delete`, { itemId })
-    .then(res => {
-      dispatch(deleteItem(res.data));
-    })
-    .catch(err => console.log(err));
+  console.log(
+    'deleteTheItem: user, itemId, from signedInCart in store',
+    userId,
+    itemId
+  );
+  //remove item from localStorage
+  let prods = [];
+  prods = JSON.parse(localStorage.getItem('products')).filter(
+    id => id !== itemId
+  );
+  console.log(prods, 'prods after delete');
+  localStorage.setItem('products', JSON.stringify(prods));
+  if (!userId) deleteItem(itemId);
+  if (typeof userId === 'number')
+    axios
+      .post(`/api/users/${userId}/item/delete`, { itemId })
+      .then(res => {
+        //remove item from store
+        dispatch(deleteItem(res.data));
+      })
+      .catch(err => console.log(err));
 };
+
 export const changedOrder = (userId, orderInfo) => dispatch =>
   axios
     .put(`api/users/${userId}/order`, orderInfo)
     .then(res => dispatch(changeOrder(res.data)))
     .catch(err => console.log(err));
 
-export default (state = {}, action) => {
+export default (state = [], action) => {
+  console.log('action', action);
   switch (action.type) {
     case ADD_ITEM_TO_CART:
+      return action.item;
+    case ADD_ITEM_TO_GUEST_CART:
       return action.item;
     case USER_ORDER:
       return action.items;
     case DELETE_ITEM:
       return state.filter(item => item.productId !== action.id);
     case CHANGE_ORDER:
-      return state;
+      return state.map(it => {
+        // if (it.id === action.item.id) it = action.item;
+        // return it;
+        ///CHECK THIS OUT  BELOWWWWWW THE RETURN
+        return state;
+      });
+    // return action.item;
     default:
       return state;
   }
